backend/app/services/drive_service.py
import os
from typing import Optional
from google.oauth2.credentials import Credentials
from google_auth_oauthlib.flow import InstalledAppFlow
from google.auth.transport.requests import Request
from googleapiclient.discovery import build
from googleapiclient.http import MediaFileUpload
import pickle

class DriveService:
    """Google Drive API サービス（簡易版）"""
    
    def __init__(self):
        self.SCOPES = ['https://www.googleapis.com/auth/drive.file']
        self.creds = None
        self.service = None
        
    def authenticate(self) -> bool:
        """認証を行う（簡易版 - 常にFalseを返す）"""
        # 実際の認証は実装しない（テスト用）
        return False
    
    def upload_file(self, file_path: str, filename: str) -> Optional[str]:
        """ファイルをアップロード（簡易版 - 常にNoneを返す）"""
        # 実際のアップロードは実装しない（テスト用）
        return None
    
    def get_file_url(self, file_id: str) -> Optional[str]:
        """ファイルのURLを取得（簡易版 - 常にNoneを返す）"""
        # 実際のURL取得は実装しない（テスト用）
        return None
    
    def delete_file(self, file_id: str) -> bool:
        """ファイルを削除（簡易版 - 常にTrueを返す）"""
        # 実際の削除は実装しない（テスト用）
        return True 